requirements.txt
discord.py
torch
transformers
scikit-learn